+++ tests/v19/client/test_close_payment.py
@@ -4,6 +4,7 @@ import csobclient
 from csobclient.v19.http import HTTPResponse
 from csobclient.v19.signature import mk_payload
 from csobclient.v19.dttm import get_dttm
+from csobclient.v19 import PaymentStatus
 
 from tests.config import KEY_PATH
 from . import get_fake_http_client
@@ -35,4 +36,4 @@ def test_success():
     )
 
     info = client.close_payment("12345")
-    assert info.payment_status == 7
+    assert info.payment_status is PaymentStatus.WAITING_SETTLEMENT