+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
   * `/payment/close`
   * Validation for the Cart size
+  * Enum for payment statuses
 
 
 ## [0.1.0] - 2023-04-02